hw 5

+++ src/services/contacts.js
@@ -1,9 +1,9 @@
 import { contactsCollection } from "../models/contacts.js";
 
-export const getAllContacts = async (query) => {
+export const getAllContacts = async (query, userId) => {
     const { page = 1, perPage = 10, sortBy, sortOrder = 'asc', type, isFavourite } = query;
 
-    const filter = {};
+    const filter = {userId};
     if (query.type) {
         filter.contactType = query.type;
     }
@@ -38,8 +38,8 @@ export const getAllContacts = async (query) => {
 };
 
 
-export const getContactById = async (contactId) => {
-    const contact = await contactsCollection.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+    const contact = await contactsCollection.findOne({ _id: contactId, userId });
     return contact;
 };
 
@@ -48,14 +48,14 @@ export const createContact = async (data) => {
     return newContact;
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await contactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+    const contact = await contactsCollection.findOneAndDelete({ _id: contactId, userId });
     return contact;
 };
 
-export const updateContactById = async (contactId, payload, options = {}) => {
+export const updateContactById = async (contactId, payload, userId, options = {}) => {
     const rawResult = await contactsCollection.findOneAndUpdate(
-        { _id: contactId },
+        { _id: contactId, userId },
         { $set: payload },
         {
             new: true,
